# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Backend API
  description: Secure RESTful API for multi-user task management with JWT authentication
  version: 1.0.0
  contact:
    name: Backend API Team

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.yourdomain.com
    description: Production server

security:
  - bearerAuth: []

tags:
  - name: tasks
    description: Task management operations

paths:
  /api/{user_id}/tasks:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'

    get:
      tags: [tasks]
      summary: List all tasks for authenticated user
      description: |
        Retrieve all tasks belonging to the authenticated user.
        Returns empty array if user has no tasks.
        Tasks sorted by creation date (newest first).
      operationId: listTasks
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
              examples:
                withTasks:
                  summary: User with tasks
                  value:
                    data:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        user_id: 123e4567-e89b-12d3-a456-426614174000
                        title: Buy groceries
                        is_completed: false
                        completed_at: null
                        created_at: '2026-01-14T10:00:00Z'
                        updated_at: '2026-01-14T10:00:00Z'
                      - id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                        user_id: 123e4567-e89b-12d3-a456-426614174000
                        title: Finish project report
                        is_completed: true
                        completed_at: '2026-01-14T11:30:00Z'
                        created_at: '2026-01-13T09:00:00Z'
                        updated_at: '2026-01-14T11:30:00Z'
                    meta:
                      timestamp: '2026-01-14T12:00:00Z'
                      count: 2
                emptyList:
                  summary: User with no tasks
                  value:
                    data: []
                    meta:
                      timestamp: '2026-01-14T12:00:00Z'
                      count: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [tasks]
      summary: Create a new task
      description: |
        Create a new task for the authenticated user.
        Task will be created with is_completed=false by default.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              validTask:
                summary: Valid task creation
                value:
                  title: Buy groceries
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  user_id: 123e4567-e89b-12d3-a456-426614174000
                  title: Buy groceries
                  is_completed: false
                  completed_at: null
                  created_at: '2026-01-14T10:00:00Z'
                  updated_at: '2026-01-14T10:00:00Z'
                meta:
                  timestamp: '2026-01-14T10:00:00Z'
                  request_id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
      - $ref: '#/components/parameters/TaskIdPath'

    get:
      tags: [tasks]
      summary: Get a specific task
      description: Retrieve a single task by ID for the authenticated user
      operationId: getTask
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  user_id: 123e4567-e89b-12d3-a456-426614174000
                  title: Buy groceries
                  is_completed: false
                  completed_at: null
                  created_at: '2026-01-14T10:00:00Z'
                  updated_at: '2026-01-14T10:00:00Z'
                meta:
                  timestamp: '2026-01-14T12:00:00Z'
                  request_id: b2c3d4e5-f6g7-8901-bcde-f12345678901
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [tasks]
      summary: Update a task
      description: |
        Update an existing task's title or completion status.
        Partial updates supported (provide only fields to update).
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            examples:
              updateTitle:
                summary: Update title only
                value:
                  title: Buy organic groceries
              updateCompletion:
                summary: Mark complete
                value:
                  is_completed: true
              updateBoth:
                summary: Update both
                value:
                  title: Buy organic groceries
                  is_completed: true
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  user_id: 123e4567-e89b-12d3-a456-426614174000
                  title: Buy organic groceries
                  is_completed: true
                  completed_at: '2026-01-14T12:30:00Z'
                  created_at: '2026-01-14T10:00:00Z'
                  updated_at: '2026-01-14T12:30:00Z'
                meta:
                  timestamp: '2026-01-14T12:30:00Z'
                  request_id: c3d4e5f6-g7h8-9012-cdef-123456789012
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [tasks]
      summary: Delete a task
      description: Permanently delete a task (hard delete, not recoverable)
      operationId: deleteTask
      responses:
        '204':
          description: Task deleted successfully (no content returned)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
      - $ref: '#/components/parameters/TaskIdPath'

    patch:
      tags: [tasks]
      summary: Mark task as complete
      description: |
        Convenience endpoint to mark a task as complete.
        Idempotent operation (calling on already-complete task succeeds).
      operationId: completeTask
      responses:
        '200':
          description: Task marked as complete successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              example:
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  user_id: 123e4567-e89b-12d3-a456-426614174000
                  title: Buy groceries
                  is_completed: true
                  completed_at: '2026-01-14T12:30:00Z'
                  created_at: '2026-01-14T10:00:00Z'
                  updated_at: '2026-01-14T12:30:00Z'
                meta:
                  timestamp: '2026-01-14T12:30:00Z'
                  request_id: d4e5f6g7-h8i9-0123-defg-234567890123
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication system (Better Auth)

  parameters:
    UserIdPath:
      name: user_id
      in: path
      required: true
      description: UUID of the user (must match JWT user_id claim)
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000

    TaskIdPath:
      name: id
      in: path
      required: true
      description: UUID of the task
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (non-empty, max 500 characters)
          example: Buy groceries

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated task title
          example: Buy organic groceries
        is_completed:
          type: boolean
          description: Completion status
          example: true

    TaskResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Task'
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
              example: '2026-01-14T12:00:00Z'
            request_id:
              type: string
              format: uuid
              example: a1b2c3d4-e5f6-7890-abcd-ef1234567890

    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        user_id:
          type: string
          format: uuid
          description: Owner user identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        title:
          type: string
          maxLength: 500
          description: Task title
          example: Buy groceries
        is_completed:
          type: boolean
          description: Completion status
          example: false
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when task was completed (null if not completed)
          example: null
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: '2026-01-14T10:00:00Z'
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2026-01-14T10:00:00Z'

    TaskListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
              example: '2026-01-14T12:00:00Z'
            count:
              type: integer
              description: Number of tasks returned
              example: 2

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: Error code
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - BAD_REQUEST
                - INTERNAL_SERVER_ERROR
                - SERVICE_UNAVAILABLE
              example: UNAUTHORIZED
            message:
              type: string
              description: Human-readable error message
              example: Invalid or expired token
            details:
              type: object
              nullable: true
              description: Additional error details (e.g., validation errors)
              example:
                field: title
                issue: must not be empty
            timestamp:
              type: string
              format: date-time
              example: '2026-01-14T12:00:00Z'
            request_id:
              type: string
              format: uuid
              example: e5f6g7h8-i9j0-1234-efgh-345678901234

  responses:
    Unauthorized:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
              details: null
              timestamp: '2026-01-14T12:00:00Z'
              request_id: e5f6g7h8-i9j0-1234-efgh-345678901234

    Forbidden:
      description: Valid JWT but user_id in URL doesn't match JWT user_id claim
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Access denied - user_id mismatch
              details: null
              timestamp: '2026-01-14T12:00:00Z'
              request_id: f6g7h8i9-j0k1-2345-fghi-456789012345

    NotFound:
      description: Task not found or doesn't belong to authenticated user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Task not found
              details: null
              timestamp: '2026-01-14T12:00:00Z'
              request_id: g7h8i9j0-k1l2-3456-ghij-567890123456

    BadRequest:
      description: Invalid request data (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: Validation error
              details:
                field: title
                issue: must not be empty
              timestamp: '2026-01-14T12:00:00Z'
              request_id: h8i9j0k1-l2m3-4567-hijk-678901234567